requirements.txt
Flask==2.3.2
Werkzeug>=2.3.3
Jinja2>=3.1.2
itsdangerous>=2.1.2
click>=8.1.3
blinker>=1.6.2
gunicorn
psycopg2-binary

